% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_plotting.R
\name{add.distr}
\alias{add.distr}
\title{add.distr}
\usage{
add.distr(..., cols = "blue", test.cols = "red")
}
\arguments{
\item{...}{One or more distributions or objects of class hint.test.}

\item{cols}{A character string vector naming the colours of the distributions. If length(cols) is less than the number of distributions, the colours will be recycled. Defaults to "blue".}

\item{test.cols}{A character string vector naming the colours to use for the regions in which the cumulative probability of the hypothesis test was derived (if it exists). If length(test.cols) is less than the number of distributions, the colours will be recycled. Defaults to "red".}
}
\value{
Plots to the current device.
}
\description{
This function will add one or more distributions or hypothesis tests to an existing plot.
}
