% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_main.R
\name{print.hint.test}
\alias{print.hint.test}
\title{print.hint.test}
\usage{
\method{print}{hint.test}(x, ...)
}
\arguments{
\item{x}{An object of class `hint.test`.}

\item{...}{Additional arguments to be passed to `print`.}
}
\value{
Prints output to the console.
}
\description{
Prints the resuls of `hint.test`.
}
