% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_plotting.R
\name{plot.hint.test}
\alias{plot.hint.test}
\title{plot.hint.test}
\usage{
\method{plot}{hint.test}(x, ...)
}
\arguments{
\item{x}{An object of class `hint.test`.}

\item{...}{Additional arguments to be passed to `plot`.}
}
\value{
Plots to the current device.
}
\description{
This function visualises the results of a Hypergeometric Intersection test.
}
\details{
Plots the relevant Hypergeometric Intersection distribution as a segment plot, and highlights the region where the observed statistic falls, i.e. the region from which the probability is computed (two.sided tests are visualised in one tail, the one with the smallest density). This can be especially useful for pedagogical purposes.
}
